feat: more pg-typed awesomeness

+++ packages/pg-typed/src/index.ts
@@ -1,3 +1,4 @@
+import assertNever from 'assert-never';
 import {SQLQuery, Queryable} from '@databases/pg';
 import {
   bulkInsert,
@@ -76,16 +77,16 @@ export interface OrderedSelectQuery<TRecord> extends SelectQuery<TRecord> {
 
 class FieldQuery<T> {
   protected readonly __query: (
-    columnName: string,
+    columnName: SQLQuery,
     sql: Queryable['sql'],
-    toValue: (columnName: string, value: unknown) => unknown,
-  ) => SQLQuery;
+    toValue: (value: unknown) => unknown,
+  ) => SQLQuery | FieldQuery<T>;
   constructor(
     query: (
-      columnName: string,
+      columnName: SQLQuery,
       sql: Queryable['sql'],
-      toValue: (columnName: string, value: unknown) => unknown,
-    ) => SQLQuery,
+      toValue: (value: unknown) => unknown,
+    ) => SQLQuery | FieldQuery<T>,
   ) {
     this.__query = query;
   }
@@ -95,27 +96,30 @@ class FieldQuery<T> {
     );
   }
   static query<T>(
-    columnName: string,
+    columnName: SQLQuery,
     q: FieldQuery<T> | unknown,
     sql: Queryable['sql'],
-    toValue: (columnName: string, value: unknown) => unknown,
+    toValue: (value: unknown) => unknown,
   ): SQLQuery {
     if (q === null) {
-      return sql`${sql.ident(columnName)} IS NULL`;
+      return sql`${columnName} IS NULL`;
     }
     if (q && q instanceof FieldQuery) {
-      return q.__query(columnName, sql, toValue);
+      const result = q.__query(columnName, sql, toValue);
+      if (result instanceof FieldQuery) {
+        return FieldQuery.query(columnName, result, sql, toValue);
+      }
+      return result;
     }
 
-    return sql`${sql.ident(columnName)} = ${toValue(columnName, q)}`;
+    return sql`${columnName} = ${toValue(q)}`;
   }
 }
 
-export type {FieldQuery};
+const FALSE_FIELD_QUERY = new FieldQuery<any>((_columnName, sql) => sql`FALSE`);
+const TRUE_FIELD_QUERY = new FieldQuery<any>((_columnName, sql) => sql`TRUE`);
 
-export type WhereCondition<TRecord> = Partial<
-  {readonly [key in keyof TRecord]: TRecord[key] | FieldQuery<TRecord[key]>}
->;
+export type {FieldQuery};
 
 export function anyOf<T>(values: {
   [Symbol.iterator](): IterableIterator<T | FieldQuery<T>>;
@@ -139,14 +143,12 @@ export function anyOf<T>(values: {
     parts.push(
       new FieldQuery<T>(
         (columnName, sql, toValue) =>
-          sql`${sql.ident(columnName)} = ANY(${valuesArray.map((v) =>
-            toValue(columnName, v),
-          )})`,
+          sql`${columnName} = ANY(${valuesArray.map((v) => toValue(v))})`,
       ),
     );
   }
   if (parts.length === 0) {
-    return new FieldQuery<T>((_columnName, sql) => sql`FALSE`);
+    return FALSE_FIELD_QUERY;
   }
   if (parts.length === 1) {
     return parts[0];
@@ -160,16 +162,40 @@ export function anyOf<T>(values: {
   );
 }
 
+export function allOf<T>(values: {
+  [Symbol.iterator](): IterableIterator<FieldQuery<T>>;
+}) {
+  for (const q of values) {
+    if (q === FALSE_FIELD_QUERY) {
+      return FALSE_FIELD_QUERY;
+    }
+  }
+  return new FieldQuery<T>(
+    (columnName, sql, toValue) =>
+      sql`(${sql.join(
+        [...values].map((p) => FieldQuery.query(columnName, p, sql, toValue)),
+        ' AND ',
+      )})`,
+  );
+}
+
 export function not<T>(value: T | FieldQuery<T>) {
+  if (value === TRUE_FIELD_QUERY) {
+    return FALSE_FIELD_QUERY;
+  } else if (value === FALSE_FIELD_QUERY) {
+    return TRUE_FIELD_QUERY;
+  }
   return new FieldQuery<T>(
     (columnName, sql, toValue) =>
       sql`NOT (${FieldQuery.query(columnName, value, sql, toValue)})`,
   );
 }
 
-function internalInQueryResults(query: (sql: Queryable['sql']) => SQLQuery) {
+function internalInQueryResults(
+  query: (sql: Queryable['sql']) => SQLQuery | FieldQuery<any>,
+) {
   return new FieldQuery<any>(
-    (columnName, sql) => sql`${sql.ident(columnName)} IN (${query(sql)})`,
+    (columnName, sql) => sql`${columnName} IN (${query(sql)})`,
   );
 }
 
@@ -179,16 +205,108 @@ export function inQueryResults(query: SQLQuery) {
 
 export function lessThan<T>(value: T) {
   return new FieldQuery<T>(
-    (columnName, sql, toValue) =>
-      sql`${sql.ident(columnName)} < ${toValue(columnName, value)}`,
+    (columnName, sql, toValue) => sql`${columnName} < ${toValue(value)}`,
   );
 }
 
 export function greaterThan<T>(value: T) {
   return new FieldQuery<T>(
-    (columnName, sql, toValue) =>
-      sql`${sql.ident(columnName)} > ${toValue(columnName, value)}`,
+    (columnName, sql, toValue) => sql`${columnName} > ${toValue(value)}`,
   );
+}
+
+export function jsonPath(
+  path: readonly string[],
+  query: any | FieldQuery<any>,
+) {
+  return new FieldQuery<any>((columnName, sql, toValue) =>
+    FieldQuery.query(sql`${columnName}#>${path}`, query, sql, toValue),
+  );
+}
+
+export function caseInsensitive(query: string | FieldQuery<string>) {
+  return new FieldQuery<string>((columnName, sql, toValue) =>
+    FieldQuery.query(sql`LOWER(${columnName})`, query, sql, (v) =>
+      `${toValue(v)}`.toLowerCase(),
+    ),
+  );
+}
+
+class WhereCombinedCondition<TRecord> {
+  protected readonly __conditions: readonly WhereCondition<TRecord>[];
+  protected readonly __combiner: 'AND' | 'OR';
+  constructor(
+    conditions: readonly WhereCondition<TRecord>[],
+    combiner: 'AND' | 'OR',
+  ) {
+    this.__conditions = conditions;
+    this.__combiner = combiner;
+  }
+  static query<T>(
+    q: WhereCombinedCondition<T> | WhereCondition<T>,
+    sql: Queryable['sql'],
+    toValue: (columnName: string, value: unknown) => unknown,
+  ): SQLQuery | 'TRUE' | 'FALSE' {
+    if (q instanceof WhereCombinedCondition) {
+      const conditions = q.__conditions.map((c) =>
+        WhereCombinedCondition.query(c, sql, toValue),
+      );
+      const significantConditions = conditions.filter(
+        (v: SQLQuery | 'TRUE' | 'FALSE'): v is SQLQuery =>
+          v !== 'TRUE' && v !== 'FALSE',
+      );
+      switch (q.__combiner) {
+        case 'AND':
+          if (conditions.some((c) => c === 'FALSE')) {
+            return 'FALSE';
+          }
+          return sql`(${sql.join(significantConditions, sql` AND `)})`;
+        case 'OR':
+          if (conditions.some((c) => c === 'TRUE')) {
+            return 'TRUE';
+          }
+          return sql`(${sql.join(significantConditions, sql` OR `)})`;
+        default:
+          return assertNever(q.__combiner);
+      }
+    }
+
+    const entries = Object.entries(q).filter(
+      (row) => row[1] !== undefined && row[1] !== TRUE_FIELD_QUERY,
+    );
+    if (entries.length === 0) {
+      return 'TRUE';
+    }
+    if (entries.some(([_column, value]) => value === FALSE_FIELD_QUERY)) {
+      return 'FALSE';
+    }
+    return sql.join(
+      entries.map(([columnName, value]) =>
+        FieldQuery.query(sql.ident(columnName), value, sql, (v) =>
+          toValue(columnName, v),
+        ),
+      ),
+      sql` AND `,
+    );
+  }
+}
+export type {WhereCombinedCondition};
+
+export type WhereCondition<TRecord> =
+  | Partial<
+      {readonly [key in keyof TRecord]: TRecord[key] | FieldQuery<TRecord[key]>}
+    >
+  | WhereCombinedCondition<TRecord>;
+
+export function and<TRecord>(
+  ...conditions: readonly WhereCondition<TRecord>[]
+): WhereCondition<TRecord> {
+  return new WhereCombinedCondition(conditions, 'AND');
+}
+export function or<TRecord>(
+  ...conditions: readonly WhereCondition<TRecord>[]
+): WhereCondition<TRecord> {
+  return new WhereCombinedCondition(conditions, 'OR');
 }
 
 interface SelectQueryOptions {
@@ -358,31 +476,6 @@ type BulkInsertRecord<
   TKey extends keyof TInsertParameters,
 > = BulkRecord<TInsertParameters, BulkInsertFields<TInsertParameters, TKey>>;
 
-function rowToCondition<TRecord>(
-  row: WhereCondition<TRecord>,
-  sql: Queryable['sql'],
-  toValue: (columnName: string, value: unknown) => unknown,
-): SQLQuery | null {
-  const entries = Object.entries(row).filter((row) => row[1] !== undefined);
-  if (entries.length === 0) {
-    return null;
-  }
-  return sql.join(
-    entries.map(([columnName, value]) =>
-      FieldQuery.query(columnName, value, sql, toValue),
-    ),
-    sql` AND `,
-  );
-}
-function rowToWhere<TRecord>(
-  row: WhereCondition<TRecord>,
-  sql: Queryable['sql'],
-  toValue: (columnName: string, value: unknown) => unknown,
-): SQLQuery {
-  const condition = rowToCondition(row, sql, toValue);
-  return condition ? sql`WHERE ${condition}` : sql``;
-}
-
 type BulkOperationOptionsBase<
   TColumnName extends string | number | symbol,
   TInsertColumnName extends string | number | symbol,
@@ -463,8 +556,12 @@ class Table<TRecord, TInsertParameters> {
       TInsertParameters,
       TColumnsToInsert[number]
     >[];
-  }) {
-    await bulkInsert<keyof TInsertParameters>({
+  }): Promise<TRecord[]> {
+    if (records.length === 0) {
+      return [];
+    }
+    const {sql} = this._underlyingDb;
+    return await bulkInsert<keyof TInsertParameters>({
       ...this._getBulkOperationOptions(),
       columnsToInsert: [
         ...new Set([
@@ -473,6 +570,7 @@ class Table<TRecord, TInsertParameters> {
         ]),
       ].sort(),
       records,
+      returning: sql`${this.tableId}.*`,
     });
   }
 
@@ -487,12 +585,14 @@ class Table<TRecord, TInsertParameters> {
     >[];
   }): UnorderedSelectQuery<TRecord> {
     const bulkOperationOptions = this._getBulkOperationOptions();
-    return this.findUntyped(
-      bulkCondition({
-        ...bulkOperationOptions,
-        whereColumnNames,
-        whereConditions,
-      }),
+    return this._findUntyped(
+      whereConditions.length
+        ? bulkCondition({
+            ...bulkOperationOptions,
+            whereColumnNames,
+            whereConditions,
+          })
+        : 'FALSE',
     );
   }
 
@@ -510,12 +610,17 @@ class Table<TRecord, TInsertParameters> {
       readonly where: BulkRecord<TRecord, TWhereColumns[number]>;
       readonly set: BulkRecord<TRecord, TSetColumns[number]>;
     }[];
-  }) {
-    await bulkUpdate<TWhereColumns[number], TSetColumns[number]>({
+  }): Promise<TRecord[]> {
+    if (updates.length === 0) {
+      return [];
+    }
+    const {sql} = this._underlyingDb;
+    return await bulkUpdate<TWhereColumns[number], TSetColumns[number]>({
       ...this._getBulkOperationOptions(),
       whereColumnNames,
       setColumnNames,
       updates,
+      returning: sql`${this.tableId}.*`,
     });
   }
 
@@ -531,6 +636,9 @@ class Table<TRecord, TInsertParameters> {
       TWhereColumns[number]
     >[];
   }) {
+    if (whereConditions.length === 0) {
+      return;
+    }
     await bulkDelete<TWhereColumns[number]>({
       ...this._getBulkOperationOptions(),
       whereColumnNames,
@@ -605,7 +713,18 @@ class Table<TRecord, TInsertParameters> {
   }
 
   async insertOrUpdate<TRecordsToInsert extends readonly TInsertParameters[]>(
-    conflictKeys: readonly [keyof TRecord, ...(keyof TRecord)[]],
+    options:
+      | readonly [keyof TRecord, ...(keyof TRecord)[]]
+      | {
+          onConflict: readonly [keyof TRecord, ...(keyof TRecord)[]];
+          set?: readonly [keyof TRecord, ...(keyof TRecord)[]];
+          doNotSet?: undefined;
+        }
+      | {
+          onConflict: readonly [keyof TRecord, ...(keyof TRecord)[]];
+          set?: undefined;
+          doNotSet?: readonly [keyof TRecord, ...(keyof TRecord)[]];
+        },
     ...rows: keyof TRecordsToInsert[number] extends keyof TInsertParameters
       ? TRecordsToInsert
       : readonly ({[key in keyof TInsertParameters]: TInsertParameters[key]} &
@@ -616,20 +735,36 @@ class Table<TRecord, TInsertParameters> {
             >]: never;
           })[]
   ): Promise<{-readonly [key in keyof TRecordsToInsert]: TRecord}> {
+    const getOption = (
+      k: 'onConflict' | 'set' | 'doNotSet',
+    ): readonly (keyof TRecord)[] | undefined => {
+      return Array.isArray(options) ? undefined : (options as any)[k];
+    };
+    const conflictKeys =
+      getOption('onConflict') ?? (options as readonly (keyof TRecord)[]);
+    const set = getOption('set');
+    const doNotSet = getOption('doNotSet');
+
     const {sql} = this._underlyingDb;
-    return this._insert(
-      (columnNames) =>
-        sql`ON CONFLICT (${sql.join(
-          conflictKeys.map((k) => sql.ident(k)),
-          sql`, `,
-        )}) DO UPDATE SET ${sql.join(
-          columnNames.map(
-            (key) => sql`${sql.ident(key)}=EXCLUDED.${sql.ident(key)}`,
-          ),
-          sql`, `,
-        )}`,
-      ...rows,
-    ) as any;
+    return this._insert((columnNames) => {
+      let updateKeys: readonly (string | number | symbol)[] = columnNames;
+      if (set) {
+        updateKeys = set;
+      }
+      if (doNotSet) {
+        const keysNotToSet = new Set<string | number | symbol>(doNotSet);
+        updateKeys = updateKeys.filter((key) => !keysNotToSet.has(key));
+      }
+      return sql`ON CONFLICT (${sql.join(
+        conflictKeys.map((k) => sql.ident(k)),
+        sql`, `,
+      )}) DO UPDATE SET ${sql.join(
+        updateKeys.map(
+          (key) => sql`${sql.ident(key)}=EXCLUDED.${sql.ident(key)}`,
+        ),
+        sql`, `,
+      )}`;
+    }, ...rows) as any;
   }
 
   async insertOrIgnore<TRecordsToInsert extends readonly TInsertParameters[]>(
@@ -654,7 +789,14 @@ class Table<TRecord, TInsertParameters> {
     updateValues: Partial<TRecord>,
   ): Promise<TRecord[]> {
     const {sql} = this._underlyingDb;
-    const where = rowToWhere(whereValues, sql, this._value);
+    const whereConditions = WhereCombinedCondition.query(
+      whereValues,
+      sql,
+      this._value,
+    );
+    if (whereConditions === `FALSE`) {
+      return [];
+    }
     const setClause = sql.join(
       Object.entries(updateValues).map(([columnName, value]) => {
         return sql`${sql.ident(columnName)} = ${this._value(
@@ -664,15 +806,31 @@ class Table<TRecord, TInsertParameters> {
       }),
       sql`, `,
     );
-    return await this.untypedQuery(
-      sql`UPDATE ${this.tableId} SET ${setClause} ${where} RETURNING *`,
-    );
+    if (whereConditions === 'TRUE') {
+      return await this.untypedQuery(
+        sql`UPDATE ${this.tableId} SET ${setClause} RETURNING *`,
+      );
+    } else {
+      return await this.untypedQuery(
+        sql`UPDATE ${this.tableId} SET ${setClause} WHERE ${whereConditions} RETURNING *`,
+      );
+    }
   }
 
   async delete(whereValues: WhereCondition<TRecord>): Promise<void> {
     const {sql} = this._underlyingDb;
-    const where = rowToWhere(whereValues, sql, this._value);
-    await this.untypedQuery(sql`DELETE FROM ${this.tableId} ${where}`);
+    const whereConditions = WhereCombinedCondition.query(
+      whereValues,
+      sql,
+      this._value,
+    );
+    if (whereConditions === 'TRUE') {
+      await this.untypedQuery(sql`DELETE FROM ${this.tableId}`);
+    } else if (whereConditions !== 'FALSE') {
+      await this.untypedQuery(
+        sql`DELETE FROM ${this.tableId} WHERE ${whereConditions}`,
+      );
+    }
   }
 
   /**
@@ -684,7 +842,9 @@ class Table<TRecord, TInsertParameters> {
     return this.find(whereValues);
   }
 
-  findUntyped(whereCondition: SQLQuery | null): UnorderedSelectQuery<TRecord> {
+  private _findUntyped(
+    whereCondition: SQLQuery | 'TRUE' | 'FALSE',
+  ): UnorderedSelectQuery<TRecord> {
     const {sql} = this._underlyingDb;
     return new SelectQueryImplementation(
       this.tableName,
@@ -694,6 +854,7 @@ class Table<TRecord, TInsertParameters> {
         limit: limitCount,
         distinctColumnNames,
       }) => {
+        if (whereCondition === 'FALSE') return [];
         return await this._underlyingDb.query(
           sql.join(
             [
@@ -711,7 +872,7 @@ class Table<TRecord, TInsertParameters> {
                   )
                 : sql`*`,
               sql`FROM ${this.tableId}`,
-              whereCondition ? sql`WHERE ${whereCondition}` : null,
+              whereCondition !== 'TRUE' ? sql`WHERE ${whereCondition}` : null,
               orderByQueries.length
                 ? sql`ORDER BY ${sql.join(
                     orderByQueries.map((q) =>
@@ -735,7 +896,9 @@ class Table<TRecord, TInsertParameters> {
     whereValues: WhereCondition<TRecord> = {},
   ): UnorderedSelectQuery<TRecord> {
     const {sql} = this._underlyingDb;
-    return this.findUntyped(rowToCondition(whereValues, sql, this._value));
+    return this._findUntyped(
+      WhereCombinedCondition.query(whereValues, sql, this._value),
+    );
   }
 
   /**
@@ -758,11 +921,24 @@ class Table<TRecord, TInsertParameters> {
 
   async count(whereValues: WhereCondition<TRecord> = {}): Promise<number> {
     const {sql} = this._underlyingDb;
-    const where = rowToWhere(whereValues, sql, this._value);
-    const [result] = await this._underlyingDb.query(
-      sql`SELECT count(*) AS count FROM ${this.tableId} ${where}`,
+    const whereCondition = WhereCombinedCondition.query(
+      whereValues,
+      sql,
+      this._value,
     );
-    return parseInt(result.count, 10);
+    if (whereCondition === `FALSE`) {
+      return 0;
+    } else if (whereCondition === `TRUE`) {
+      const [result] = await this._underlyingDb.query(
+        sql`SELECT count(*) AS count FROM ${this.tableId}`,
+      );
+      return parseInt(result.count, 10);
+    } else {
+      const [result] = await this._underlyingDb.query(
+        sql`SELECT count(*) AS count FROM ${this.tableId} WHERE ${whereCondition}`,
+      );
+      return parseInt(result.count, 10);
+    }
   }
 
   async untypedQuery(query: SQLQuery): Promise<TRecord[]> {
@@ -832,14 +1008,25 @@ function getTable<TRecord, TInsertParameters>(
         fieldName: TKey,
         condition: WhereCondition<TRecord> = {},
       ): FieldQuery<TRecord[TKey]> =>
-        internalInQueryResults(
-          (sql) =>
-            sql`SELECT ${sql.ident(fieldName)} FROM ${
-              schemaName
-                ? sql.ident(schemaName, tableName)
-                : sql.ident(tableName)
-            } ${rowToWhere(condition, sql, serializeValue)}`,
-        ),
+        internalInQueryResults((sql) => {
+          const whereCondition = WhereCombinedCondition.query(
+            condition,
+            sql,
+            serializeValue,
+          );
+          if (whereCondition === 'FALSE') {
+            return FALSE_FIELD_QUERY;
+          }
+          const tableId = schemaName
+            ? sql.ident(schemaName, tableName)
+            : sql.ident(tableName);
+          const fieldId = sql.ident(fieldName);
+          if (whereCondition === 'TRUE') {
+            return sql`SELECT ${fieldId} FROM ${tableId}`;
+          } else {
+            return sql`SELECT ${fieldId} FROM ${tableId} WHERE ${whereCondition}`;
+          }
+        }),
     },
   );
 }
@@ -1002,7 +1189,12 @@ module.exports = Object.assign(defineTables, {
   not,
   inQueryResults,
   lessThan,
+  jsonPath,
+  caseInsensitive,
   greaterThan,
+  allOf,
+  and,
+  or,
   isNoResultFoundError,
   isMultipleResultsFoundError,
 });
